updated image mutations validation

+++ app/graphql/mutations/images/remove_image_from_comment.rb
@@ -8,8 +8,15 @@ module Mutations
       field :errors, [String], null: true
 
       def resolve(comment_id:, image_id:)
-        comment = current_user.comments.find(comment_id)
-        image = comment.images.find(image_id)
+        comment = Comment.find_by(id: comment_id)
+
+        raise GraphQL::ExecutionError, "Comment not found" unless comment.present?
+        raise GraphQL::ExecutionError, "You cannot remove image from this comment" unless comment.user_id == current_user.id
+
+        image = Image.find_by(id: image_id)
+
+        raise GraphQL::ExecutionError, "Image not found" unless image.present?
+        raise GraphQL::ExecutionError, "Comment doesn't contain this image" unless comment.images.exists?(imageable: image)
 
         image.destroy
